Выводится дата создания

+++ domain/src/main/kotlin/ru/surfstudio/standard/domain/feed/Meme.kt
@@ -2,7 +2,7 @@ package ru.surfstudio.standard.domain.feed
 
 data class Meme(
         val id: String,
-        val createdDate: Int,
+        val createdDate: Long,
         val description: String,
         val isFavorite: Boolean,
         val photoUtl: String,
